+++ R/bertinplot.R
@@ -119,16 +119,16 @@
 #' bertinplot(x, order, panel = panel.lines)
 #'
 #' # Plot with cut lines (we manually set the order here)
-#' order <- ser_permutation(c(21, 16, 19, 18, 14, 12, 20, 15,
-#'     17, 26, 13, 41,  7, 11, 5, 23, 28, 34, 31, 1, 38, 40,
-#'     3, 39,  4, 27, 24,  8, 37, 36, 25, 30, 33, 35,  2,
-#'     22, 32, 29, 10,  6,  9),
-#'     c(4, 2, 1, 6, 8, 7, 5, 3))
+#' order <- ser_permutation(c(6L, 9L, 29L, 10L, 32L, 22L, 2L, 35L,
+#'   24L, 30L, 33L, 25L, 37L, 36L, 8L, 27L, 4L, 39L, 3L, 40L, 38L,
+#'   1L, 31L, 34L, 28L, 23L, 5L, 11L, 7L, 41L, 13L, 26L, 17L, 15L,
+#'   12L, 20L, 14L, 18L, 19L, 16L, 21L),
+#'     c(4L, 2L, 1L, 6L, 7L, 8L, 5L, 3L))
 #'
 #' bertinplot(x, order, pop=FALSE)
 #' bertin_cut_line(, 4) ## horizontal line between rows 4 and 5
 #' bertin_cut_line(, 7) ## separate "Right to Life" from the rest
-#' bertin_cut_line(14, c(0, 4)) ## separate a block of large values (vertically)
+#' bertin_cut_line(18, c(0, 4)) ## separate a block of large values (vertically)
 #'
 #' # ggplot2-based plots
 #' if (require("ggplot2")) {
@@ -167,13 +167,13 @@
 #' }
 #' @export
 bertinplot <- function(x,
-  order = NULL,
-  panel.function = panel.bars,
-  highlight = TRUE,
-  row_labels = TRUE,
-  col_labels = TRUE,
-  flip_axes = TRUE,
-  ...) {
+                       order = NULL,
+                       panel.function = panel.bars,
+                       highlight = TRUE,
+                       row_labels = TRUE,
+                       col_labels = TRUE,
+                       flip_axes = TRUE,
+                       ...) {
   if (!is.matrix(x))
     stop("Argument 'x' must be a matrix.")
 
@@ -215,12 +215,9 @@ bertinplot <- function(x,
 
   ## note: Bertin switched cols and rows for his display!
   # change x and y?
-  if (flip_axes) {
+  if (flip_axes)
     x <- t(x)
-    tmp <- row_labels
-    row_labels <- col_labels
-    col_labels <- tmp
-  }
+
 
   ## highlight
   if (is.logical(highlight) && highlight)
@@ -256,17 +253,17 @@ bertinplot <- function(x,
   x <- map(x)
 
   for (variable in seq(nrow(x))) {
-    value <- x[variable, ]
-    hl <- col[variable, ]
+    value <- x[variable,]
+    hl <- col[variable,]
 
     ## handle neg. values
     if (identical(options$panel.function, panel.bars) ||
         identical(options$panel.function, panel.lines)) {
       ylim <- c(min(value, 0, na.rm = TRUE),
-        max(value, 0, na.rm = TRUE) + options$spacing)
+                max(value, 0, na.rm = TRUE) + options$spacing)
     } else{
       ylim <- c(0,
-        max(abs(value), 0.1, na.rm = TRUE))
+                max(abs(value), 0.1, na.rm = TRUE))
     }
 
     pushViewport(
@@ -301,24 +298,26 @@ bertinplot <- function(x,
   else
     0
 
-  grid.text(
-    colnames(x),
-    x = seq(ncol(x)),
-    y = nrow(x) + spacing_corr,
-    rot = 90,
-    just = "left",
-    default.units = "native",
-    gp = options$gp_labels
-  )
+  if (col_labels)
+    grid.text(
+      colnames(x),
+      x = seq(ncol(x)),
+      y = nrow(x) + spacing_corr,
+      rot = 90,
+      just = "left",
+      default.units = "native",
+      gp = options$gp_labels
+    )
 
-  grid.text(
-    rev(rownames(x)),
-    x = 1 + spacing_corr / ncol(x) / 4,
-    y = 0.5:(nrow(x) - 0.5) / nrow(x),
-    just = "left",
-    default.units = "npc",
-    gp = options$gp_labels
-  )
+  if (row_labels)
+    grid.text(
+      rev(rownames(x)),
+      x = 1 + spacing_corr / ncol(x) / 4,
+      y = 0.5:(nrow(x) - 0.5) / nrow(x),
+      just = "left",
+      default.units = "npc",
+      gp = options$gp_labels
+    )
 
   if (options$pop)
     popViewport(1)
@@ -416,8 +415,8 @@ panel.lines <- function(value, spacing, hl) {
 #' @rdname bertinplot
 #' @export
 bertin_cut_line <- function(x = NULL,
-  y = NULL,
-  col = "red") {
+                            y = NULL,
+                            col = "red") {
   if (length(x) < 2)
     x <- rep(x, 2)
   if (length(y) < 2)
@@ -425,7 +424,7 @@ bertin_cut_line <- function(x = NULL,
 
   ## find the bertin Viewport
   if (inherits(try(seekViewport("bertin"), silent = TRUE)
-    , "try-error")) {
+               , "try-error")) {
     stop("bertinplot() needs to be called with options = list(pop = FALSE) first!")
   }