+++ R/pimage.R
@@ -25,36 +25,34 @@
 #' Permutation Image Plot
 #'
 #' Provides methods for matrix shading, i.e., displaying a color image for
-#' matrix (including correlation matrices) and \code{dist} objects given an
-#' optional permutation.  The plot arranges colored rectangles to represent the
-#' matrix value. Columns and rows appear in the order in the matrix.  This
-#' visualization is also know asi a heatmap.  Implementations based on the
+#' matrix (including correlation matrices and data frames) and `dist` objects given an
+#' optional permutation. The plot arranges colored rectangles to represent the
+#' values in the matrix. This visualization is also know as a heatmap.
+#' Implementations based on the
 #' \pkg{grid} graphics engine and based n \pkg{ggplot2} are provided.
 #'
-#' Plots a matrix in its original row and column orientation. This means, in a
+#' Plots a matrix in its original row and column orientation ([image] in \pkg{stats}
+#' reverses the rows). This means, in a
 #' plot the columns become the x-coordinates and the rows the y-coordinates (in
 #' reverse order).
 #'
-#' If \code{x} is of class \code{dist} it is converted to full-storage
-#' representation before plotting.
+#' **Grid-based plot:** The viewports used for plotting are called:
+#' `"plot"`, `"image"` and `"colorkey"`.  Use \pkg{grid} functions
+#' to manipulate the plots (see Examples section).
 #'
-#' \bold{Grid-based plot:} The viewports used for plotting are called:
-#' \code{"plot"}, \code{"image"} and \code{"colorkey"}.  \emph{Note:} Since
-#' \code{pimage} uses \pkg{grid}, it should not be mixed with base R primitive
-#' plotting functions, but the appropriate functions in
-#' \code{\link[grid]{grid-package}}.
-#'
-#' \bold{ggplot2-based plot:} A ggplot2 object is returned. Colors, axis limits
+#' **ggplot2-based plot:** A ggplot2 object is returned. Colors, axis limits
 #' and other visual aspects can be added using standard ggplot2 functions
-#' (\code{labs}, \code{scale_fill_continuous}, \code{labs}, etc.).
+#' (`labs`, `scale_fill_continuous`, `labs`, etc.).
 #'
 #' @family plots
 #'
-#' @param x a matrix or an object of class \code{dist}.
-#' @param order an object of class \code{ser_permutation} or the name of a seriation method.
-#'   If \code{NULL} the order in \code{x} is plotted.
-#' @param col a list of colors used. If \code{NULL}, a gray scale is used (for
-#' matrix larger values are displayed darker and for \code{dist} smaller
+#' @param x a matrix, a data.frame, or an object of class `dist`.
+#' @param order a logical where `FALSE` means no reordering and `TRUE` applies
+#'  a permutation using the default seriation method for the type of `x`. Alternatively,
+#'  any object that can be coerced to class `ser_permutation`
+#'  can be supplied.
+#' @param col a list of colors used. If `NULL`, a gray scale is used (for
+#' matrix larger values are displayed darker and for `dist` smaller
 #' distances are darker). For matrices containing logical data, black and white
 #' is used. For matrices containing negative values a symmetric diverging color
 #' palette is used.
@@ -65,13 +63,13 @@
 #' @param key logical; add a color key? No key is available for logical
 #' matrices.
 #' @param keylab string plotted next to the color key.
-#' @param symkey logical; if \code{x} contains negative values, should the
-#' color palate be symmetric (zero is in the middle)>
+#' @param symkey logical; if `x` contains negative values, should the
+#' color palate be symmetric (zero is in the middle)?
 #' @param upper_tri,lower_tri,diag a logical indicating whether to show the
 #' upper triangle, the lower triangle or the diagonal of the (distance) matrix.
 #' @param row_labels,col_labels a logical indicating if row and column labels
-#' in \code{x} should be displayed.  If \code{NULL} then labels are displayed
-#' if the \code{x} contains the appropriate dimname and the number of labels is
+#' in `x` should be displayed.  If `NULL` then labels are displayed
+#' if the `x` contains the appropriate dimname and the number of labels is
 #' 25 or less. A character vector of the appropriate length with labels can
 #' also be supplied.
 #' @param prop logical; change the aspect ratio so cells in the image have a
@@ -79,84 +77,78 @@
 #' @param flip_axes logical; exchange rows and columns for plotting.
 #' @param reverse_columns logical; revers the order of how the columns are
 #' displayed.
-#' @param \dots if `order` is the name of a seriation method then further arguments are  passed
-#'   on to the seriatation method, otherwise they are ignored.
+#' @param \dots if `order` is the name of a seriation method then further arguments are passed
+#'   on to the seriation method, otherwise they are ignored.
 #' @param newpage,pop,gp Start plot on a new page, pop the viewports after
-#' plotting, and use the supplied \code{gpar} object (see \pkg{grid}).
+#' plotting, and use the supplied `gpar` object (see \pkg{grid}).
 #' @returns Nothing.
 #'
 #' @author Christian Buchta and Michael Hahsler
 #' @keywords hplot
 #' @examples
 #' set.seed(1234)
+#' data(iris)
+#' x <- as.matrix(iris[sample(nrow(iris), 20) , -5])
 #'
-#' ## Example: Logical Matrix
-#' x <- matrix(sample(c(FALSE, TRUE), 300, rep = TRUE), ncol = 10,
-#'   dimnames = list(1:30, LETTERS[1:10]))
-#'
-#' # Matrix for logical values. TRUE values are dark and no color key is shown. There are too many
-#' # Row labels (>25) so they are suppressed.
 #' pimage(x)
 #'
-#' # Show all labels and flip axes or reverse columns
-#' pimage(x, row_labels = TRUE, col_labels = TRUE, flip_axes = TRUE)
-#' pimage(x, row_labels = TRUE, col_labels = TRUE, reverse_columns = TRUE)
-#'
-#' # Reorder matrix, use custom colors, and add a title.
-#' pimage(x, order = seriate(x), row_labels = TRUE, col_labels = TRUE,
-#'   col = c("white", "red"), main = "Random Data (Reordered)")
-#'
-#' ## Example: Positive Matrix
-#' x <- matrix(runif(100), ncol = 10,
-#'   dimnames = list(LETTERS[1:10], paste0("X", 1:10)))
-#'
+#' # Show all labels and flip axes, reverse columns, or change colors
 #' pimage(x, prop = TRUE)
+#' pimage(x, flip_axes = TRUE)
+#' pimage(x, reverse_columns = TRUE)
+#' pimage(x, col = grays(100))
 #'
-#' ## Example: Pos/Neg. Matrix
-#' x <- matrix(rnorm(100), ncol = 10,
-#'   dimnames = list(LETTERS[1:10], paste0("X", 1:10)))
+#' # A matrix with positive and negative values
+#' x_scaled <- scale(x)
+#' pimage(x_scaled)
 #'
-#' pimage(x, prop = TRUE)
+#' # Use reordering
+#' pimage(x_scaled, order = TRUE)
+#' pimage(x_scaled, order = "Heatmap")
 #'
 #' ## Example: Distance Matrix
 #' # Show a reordered distance matrix (distances between rows).
-#' # Dark means low distance. The aspect ratio is automatically fixed to 1:1.
+#' # Dark means low distance. The aspect ratio is automatically fixed to 1:1
+#' # using prop = TRUE
 #' d <- dist(x)
-#' pimage(d,  order = seriate(d),
-#'   main = "Random Data (Distances)")
+#' pimage(d)
+#' pimage(d, order = TRUE)
 #'
 #' # Supress the upper triangle and diagonal
-#' pimage(d,  order = seriate(d), upper = FALSE, diag = FALSE,
-#'   main = "Random Data (Distances)")
+#' pimage(d, order = TRUE, upper = FALSE, diag = FALSE)
 #'
-#' # Show only distances that are smaller than 4 using limits on z.
-#' pimage(d, order = seriate(d),
-#'   main = "Random Data (Distances + Theshold)", zlim = c(0, 4))
+#' # Show only distances that are smaller than 2 using limits on z.
+#' pimage(d, order = TRUE, zlim = c(0, 3))
 #'
 #' ## Example: Correlation Matrix
 #' # we calculate correlation between rows and seriate the matrix
+#' # and seriate by converting the correlations into distances.
+#' # pimage reorders then rows and columns with c(o, o).
 #' r <- cor(t(x))
-#' r <- permute(r, seriate(r))
-#' pimage(r, upper = FALSE, diag = FALSE, zlim = c(-1, 1), reverse_columns = TRUE,
-#'   main = "Random Data (Correlation)")
+#' o <- seriate(as.dist(sqrt(1 - r)))
+#' pimage(r, order = c(o, o),
+#'   upper = FALSE, diag = FALSE,
+#'   zlim = c(-1, 1),
+#'   reverse_columns = TRUE,
+#'   main = "Correlation matrix")
 #'
 #' # Add to the plot using functions in package grid
 #' # Note: pop = FALSE allows us to manipulate viewports
 #' library("grid")
-#' pimage(x, pop = FALSE)
+#' pimage(x, order = TRUE, pop = FALSE)
 #'
 #' # available viewports are: "main", "colorkey", "plot", "image"
 #' current.vpTree()
 #'
-#' # Highlight cell column 7 (G) / row 5 (from top)/col with a red arrow starting at 5/2
+#' # Highlight cell 2/2 with a red arrow
 #' # Note: columns are x and rows are y.
 #' downViewport(name = "image")
-#' grid.lines(x = c(5, 7), y = c(2, 5), arrow = arrow(),
+#' grid.lines(x = c(1, 2), y = c(-1, 2), arrow = arrow(),
 #'   default.units = "native", gp = gpar(col = "red", lwd = 3))
 #'
-#' # add a red box around rows 15 and 16
-#' grid.rect(x = 0.5, y = 5.5, width = ncol(x), height = 2,
-#'   just = "left",
+#' # add a red box around the first 4 rows of the 2nd column
+#' grid.rect(x = 1 + .5 , y = 4 + .5, width = 1, height = 4,
+#'   hjust = 0, vjust = 1,
 #'   default.units = "native", gp = gpar(col = "red", lwd = 3, fill = NA))
 #'
 #' ## remove the viewports
@@ -174,14 +166,13 @@
 #' pushViewport(splot)
 #'
 #' seekViewport("col1_vp")
-#' o <- seriate(x)
-#' pimage(x, o, col_labels = FALSE, main = "Random Data",
+#' o <- seriate(d)
+#' pimage(x, c(o, NA), col_labels = FALSE, main = "Data",
 #'   newpage = FALSE)
 #'
 #' seekViewport("col2_vp")
 #' ## add the reordered dissimilarity matrix for rows
-#' d <- dist(x)
-#' pimage(d, o[[1]], main = "Random Data",
+#' pimage(d, o, main = "Distances",
 #'   newpage = FALSE)
 #'
 #' popViewport(0)
@@ -192,73 +183,64 @@
 #'
 #' library("ggplot2")
 #'
-#' ## Example: Logical Matrix
-#' x <- matrix(sample(c(FALSE, TRUE), 300, rep = TRUE), ncol = 10,
-#'   dimnames = list(1:30, LETTERS[1:10]))
+#' set.seed(1234)
+#' data(iris)
+#' x <- as.matrix(iris[sample(nrow(iris), 20) , -5])
 #'
-#' # Matrix for logical values. TRUE values are dark. There are too many
-#' # Row labels (>25) so they are suppressed.
 #' ggpimage(x)
 #'
-#' # Show all labels and flip axes or reverse columns
-#' ggpimage(x, flip_axes = TRUE, row_labels = TRUE, col_labels = TRUE)
-#' ggpimage(x, reverse_columns = TRUE, row_labels = TRUE, col_labels = TRUE)
+#' # Show all labels and flip axes, reverse columns
+#' ggpimage(x, prop = TRUE)
+#' ggpimage(x, flip_axes = TRUE)
+#' ggpimage(x, reverse_columns = TRUE)
+#'
+#'
+#' # A matrix with positive and negative values
+#' x_scaled <- scale(x)
+#' ggpimage(x_scaled)
+#'
+#' # Use reordering
+#' ggpimage(x_scaled, order = TRUE)
+#' ggpimage(x_scaled, order = "Heatmap")
+#'
+#' ## Example: Distance Matrix
+#' # Show a reordered distance matrix (distances between rows).
+#' # Dark means low distance. The aspect ratio is automatically fixed to 1:1
+#' # using prop = TRUE
+#' d <- dist(x)
+#' ggpimage(d)
+#' ggpimage(d, order = TRUE)
+#'
+#' # Supress the upper triangle and diagonal
+#' ggpimage(d, order = TRUE, upper = FALSE, diag = FALSE)
+#'
+#' # Show only distances that are smaller than 2 using limits on z.
+#' ggpimage(d, order = TRUE, zlim = c(0, 2))
+#'
+#' ## Example: Correlation Matrix
+#' # we calculate correlation between rows and seriate the matrix
+#' r <- cor(t(x))
+#' o <- seriate(as.dist(sqrt(1 - r)))
+#' ggpimage(r, order = c(o, o),
+#'   upper = FALSE, diag = FALSE,
+#'   zlim = c(-1, 1),
+#'   reverse_columns = TRUE) + labs(title = "Correlation matrix")
+#'
+#' ## Example: Custom themes and colors
+#' # Reorder matrix, use custom colors, add a title,
+#' # and hide colorkey.
+#' ggpimage(x) +
+#'   theme(legend.position = "none") +
+#'   labs(title = "Random Data") + xlab("Variables")
 #'
 #' # Add lines
 #' ggpimage(x) +
 #'   geom_hline(yintercept = seq(0, nrow(x)) + .5) +
 #'   geom_vline(xintercept = seq(0, ncol(x)) + .5)
 #'
-#' # Reorder matrix, use custom colors, add a title,
-#' # and hide colorkey.
-#' ggpimage(x, order = seriate(x), row_labels = TRUE, col_labels = TRUE) +
-#'   scale_fill_manual(values = c("grey90", "red")) +
-#'   theme(legend.position = "none") +
-#'   labs(title = "Random Data")
-#'
-#' ## Example: Positive Matrix
-#' x <- matrix(runif(100), ncol = 10,
-#'   dimnames = list(LETTERS[1:10], paste0("X", 1:10)))
-#'
-#' ggpimage(x, order = seriate(x)) +
-#'   labs(title = "Random Data")
-#'
-#' #' ## Example: Pos/Neg. Matrix
-#' x <- matrix(rnorm(100), ncol = 10,
-#'   dimnames = list(LETTERS[1:10], paste0("X", 1:10)))
-#'
-#' ggpimage(x, order = seriate(x)) +
-#'   labs(title = "Random Data")
-#'
-#' ## Example: Distance Matrix
-#' # Show a reordered distance matrix (distances between rows).
-#' # Dark means low distance. The aspect ratio is automatically fixed to 1:1.
-#' # The upper triangle is suppressed triangle
-#' d <- dist(x)
-#' ggpimage(d, order = seriate(d)) +
-#'   labs(title = "Random Data", subtitle = "Distances")
-#'
-#' # Hide upper triangle and diagonal
-#' ggpimage(d, order = seriate(d), upper_tri = FALSE, diag = FALSE) +
-#'   labs(title = "Random Data", subtitle = "Distances")
-#'
-#' # Show only distances that are smaller than 4 using limits on fill.
-#' ggpimage(d,  order = seriate(d), zlim = c(0, 4)) +
-#'   labs(title = "Random Data (Distances + Theshold)")
-#'
-#' ## Example: Correlation Matrix
-#' # we calculate correlation between rows and seriate the matrix
-#' r <- cor(t(x))
-#' r <- permute(r, seriate(r))
-#' ggpimage(r,  zlim = c(-1, 1), upper = FALSE, diag = FALSE, reverse_columns = TRUE) +
-#'   geom_text(aes(x = col, y = row, label = round(x, 2)), color = "black", size = 4) +
-#'   labs(title = "Random Data", subtitle = "Correlation")
-#'
-#' ## Example: Custom themes and colors
 #' # Use ggplot2 themes with theme_set
 #' old_theme <- theme_set(theme_linedraw())
-#' ggpimage(d, order = seriate(d)) +
-#'   labs(title = "Random Data (Distances)")
+#' ggpimage(d)
 #' theme_set(old_theme)
 #'
 #' # Use custom color palettes: Gray scale, Colorbrewer (provided in ggplot2) and colorspace
@@ -274,7 +256,7 @@
 #' @export
 pimage <-
   function(x,
-    order = NULL,
+    order = FALSE,
     ...)
 UseMethod("pimage")
 
@@ -283,7 +265,7 @@ UseMethod("pimage")
 #' @export
 pimage.matrix <-
   function(x,
-    order = NULL,
+    order = FALSE,
     col = NULL,
     main = "",
     xlab = "",
@@ -320,13 +302,20 @@ pimage.matrix <-
     if (is.null(col)) {
       if (is.logical(x))
         col <- c("white", "black")
-      else if (any(x < 0, na.rm = TRUE))  {
-        col <- .diverge_pal(100)
-        if (is.null(zlim) && symkey)
+      else {
+        if (!is.null(zlim)) {
+          if (min(zlim) < 0)
+            col <- .diverge_pal(100)
+          else
+            col <- .sequential_pal(100)
+        } else {
+          if (any(x < 0, na.rm = TRUE))  {
+          col <- .diverge_pal(100)
           zlim <- max(abs(range(x, na.rm = TRUE))) * c(-1, 1)
+        } else
+          col <- .sequential_pal(100)
+        }
       }
-      else
-        col <- .sequential_pal(100)
     }
 
     if (is.null(prop))