+++ NEWS.md
@@ -1,3 +1,17 @@
+# seriation 1.5.0-1 (xx/xx/2023)
+
+## New Fearures
+- pimage and permute now accept order = TRUE to perform the default seriation.
+- hmap gained parameter col_dist to define the color palette used for distance
+  matrices.
+- hmap dropped parameter showDend and gained parameter plot_margins instead.
+
+## Bug Fixes
+- pimage/ggpimage now use zlim correctly to choose the color palette.
+- BEA for matrix is not correctly registered as randomized.
+- fixed col/row_labels parameter.
+- rev() for seriations based on hclust now also reverse the dendrogram.
+
 # seriation 1.5.0 (07/19/2023)
 
 ## New Features